require the service to be running, otherwise a reload will fail
A reload can only happen if the service is running. If you are
setting up the node for the first time, the service might not yet
run and the reload will fail, still later the service will be
started and everything is fine. Nevertheless, you will have a
failure in the logs.
With this change there might be still a service start and a reload,
but both will succeed.