Make sure no forwarder is set on master/slave zone
In case bind9 is master for a zone and has a zone delegation inside the
zone, having a forwarder defined makes the resolving of the NS record
fail for some reason if the NS target is inside the zone.
Apparently bind will try to forward the resolution of the NS target and
this will fail since the master for the zone is actually itself and not
the forwarder.

Removing the forwarder will actually make bind behave correctly in all
cases:
- using the cache for the name if present in it
- resolving locally if it is master for the zone
- forwarding the request if none of the previous apply